fix: fix stubs for tests
