+++ internal/gowasm/entrypoint_js.s
@@ -1,6 +1,6 @@
 // FILE WAS GENERATED BY gowasm-gen-import, DO NOT EDIT!
 //
-// Source file: worker_js.go
+// Source file: entrypoint_js.go
 
 #include "textflag.h"
