+++ internal/gowasm/packagedb/syscall_stubs.go
@@ -2,7 +2,7 @@
 
 package packagedb
 
-func lookupPackage(pkgName string, out []byte, cb int) {
+func lookupPackage(pkgName string, out *[]byte, cb int) {
 	panic("not implemented")
 }
 func registerPackage(pkgName, version string, cb int) {
